Created scripts for custom ranges for the line charts
Basically, what this does is it force corrects min/max order while using an existing lookup object pattern that was used for the piechart. It uses filters for only the selected provider's data points, extracts the metric_value numbers, and adds a "5% padding" (the .95 multiplier) so the line doesn't touch the edges.