Added introductory procedure queries
The problem with the Bitewing one was that "bitewing" was just a small part of the many services included in the "service" column of the data. So, the query had to account for any and all possible errors/typos. Plus, it had to be filtered twice for both year and the service itself, then ordered.